+++ remoulade/message.py
@@ -19,9 +19,9 @@ import time
 import uuid
 from collections import namedtuple
 
-from .broker import get_broker
 from .composition import pipeline
 from .encoder import Encoder, JSONEncoder
+from .result import Result
 
 #: The global encoder instance.
 global_encoder = JSONEncoder()
@@ -107,35 +107,9 @@ class Message(namedtuple("Message", (
         options.update(updated_options)
         return self._replace(**attributes, options=options)
 
-    def get_result(self, *, block=False, timeout=None, raise_on_error=True, forget=False):
-        """Get the result associated with this message from a result
-        backend.
-
-        Parameters:
-          block(bool): Whether or not to block while waiting for a
-            result.
-          timeout(int): The maximum amount of time, in ms, to block
-            while waiting for a result.
-          raise_on_error(bool): raise an error if the result stored in
-            an error
-          forget(bool): if true the result is discarded from the result
-            backend
-
-        Raises:
-          RuntimeError: If there is no result backend on the default
-            broker.
-          ResultMissing: When block is False and the result isn't set.
-          ResultTimeout: When waiting for a result times out.
-          ErrorStored: When the result is an error and raise_on_error is True
-
-        Returns:
-          object: The result.
-        """
-        broker = get_broker()
-        backend = broker.get_result_backend()
-
-        return backend.get_result(self.message_id, block=block, timeout=timeout, forget=forget,
-                                  raise_on_error=raise_on_error)
+    @property
+    def result(self):
+        return Result(message_id=self.message_id)
 
     def __str__(self):
         return "%s / %s" % (self.actor_name, self.message_id)